+++ rapyer/__init__.py
@@ -1,10 +1,12 @@
 """Redis Pydantic - Pydantic models with Redis as the backend."""
 
-from rapyer.base import AtomicRedisModel
+from rapyer.base import AtomicRedisModel, get, find_redis_models
 from rapyer.init import init_rapyer, teardown_rapyer
 
 __all__ = [
     "AtomicRedisModel",
     "init_rapyer",
     "teardown_rapyer",
+    "get",
+    "find_redis_models",
 ]