+++ rapyer/init.py
@@ -1,10 +1,9 @@
 import redis.asyncio as redis_async
+from rapyer.base import REDIS_MODELS
 from redis.asyncio.client import Redis
 
-from rapyer.base import REDIS_MODELS
 
-
-def init_rapyer(redis: str | Redis = None, ttl: int = None):
+async def init_rapyer(redis: str | Redis = None, ttl: int = None):
     if isinstance(redis, str):
         redis = redis_async.from_url(redis, decode_responses=True, max_connection=20)