+++ rapyer/types/dct.py
@@ -1,7 +1,6 @@
-from typing import TypeVar, Generic, get_args, Any
+from typing import TypeVar, Generic, get_args, Any, TypeAlias
 
 from pydantic_core import core_schema
-
 from rapyer.types.base import GenericRedisType, RedisType, REDIS_DUMP_FLAG_NAME
 from rapyer.utils.redis import update_keys_in_pipeline
 
@@ -227,3 +226,6 @@ class RedisDict(dict[str, T], GenericRedisType, Generic[T]):
     @classmethod
     def schema_for_unknown(cls):
         core_schema.dict_schema(core_schema.str_schema(), core_schema.str_schema())
+
+
+RedisDictType: TypeAlias = RedisDict[T] | dict[str, T]