+++ rapyer/types/lst.py
@@ -3,8 +3,8 @@ from typing import TypeVar
 
 from pydantic_core import core_schema
 from pydantic_core.core_schema import ValidationInfo, SerializationInfo
-
 from rapyer.types.base import GenericRedisType, RedisType, REDIS_DUMP_FLAG_NAME
+from typing_extensions import TypeAlias
 
 T = TypeVar("T")
 
@@ -158,3 +158,6 @@ class RedisList(list, GenericRedisType[T]):
     @classmethod
     def schema_for_unknown(cls):
         core_schema.list_schema(core_schema.str_schema())
+
+
+RedisListType: TypeAlias = RedisList[T] | list[T]