+++ rapyer/types/string.py
@@ -1,3 +1,5 @@
+from typing import TypeAlias
+
 from rapyer.types.base import RedisType
 
 
@@ -6,3 +8,12 @@ class RedisStr(str, RedisType):
 
     def clone(self):
         return str(self)
+
+    def __iadd__(self, other):
+        new_value = self + other
+        if self.pipeline:
+            self.pipeline.json().set(self.key, self.json_path, new_value)
+        return self.__class__(new_value)
+
+
+RedisStrType: TypeAlias = RedisStr | str